automatically fail the build when there are wider implementations available
This is a workaround for a SUPERCOP issue. The benchmarks decide which
implementation is fastest based on input lenght of 1536 (called
TUNE_BYTES), but this doesn't distinguish our implementations in a
couple of important cases:

- Our NEON implementation falls back to portable for inputs below 4 KiB.
- Similarly, our AVX2 implementation falls back to SSE4.1 for inputs below 4 KiB.

This means that on a NEON-supporting or AVX2-supporting machine, whether
we benchmark the correct implementation is kind of random. To work
around this, take advantage of the fact that SUPERCOP compiles us with
-march=native to automatically fail the build for all implementations
that aren't the widest supported implementation on a given platform.